requirements.txt
django==4.2.4
crispy-bootstrap5==0.7
django-crispy-forms==2.0
Pillow==10.0.0